Refactored State Changes and DeleteModel is broken
